refactor: split TaxEstimates into focused components
Decompose the monolithic TaxEstimates component into smaller, focused modules:
- TaxEstimatesViewContent: handles onboarding status routing
- TaxEstimatesViewHeader: year picker and menu actions
- TaxEstimatesOnboardedViewContent: tab navigation between estimates/payments
- taxEstimatesTypes: shared type definitions
