+++ src/views/TaxEstimates/taxEstimatesOnboardedViewContent.scss
@@ -1,0 +1,3 @@
+.Layer__TaxEstimates__TaxBannerWrapper {
+  inline-size: min(100%, 1406px);
+}
