+++ src/views/TaxEstimates/taxEstimatesTypes.ts
@@ -1,0 +1,5 @@
+import type { TaxBannerReviewPayload } from '@components/TaxDetails/TaxBanner'
+
+export type TaxEstimatesViewProps = {
+  onTaxBannerReviewClick?: (payload: TaxBannerReviewPayload) => void
+}
